[kernel][vmm] Reimplement VMM with level-dependent functions
The previous VMM implementation was design for 2-level translation
and contained types and functions which are not scalable for 4-level
translation models (e.g page_desc_t or _vmm_look_ptable).

The new design brings a concept of ptables' levels. This allows to
get rid of strong-typed functions working with ptables at any level.